% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_doc_md.R
\name{get_oida_metadata}
\alias{get_oida_metadata}
\title{Get OIDA document metadata}
\usage{
get_oida_metadata(file_names)
}
\arguments{
\item{file_names}{Eight character document identifier in the format of
"AbCd1234" where the first four characters are letters and the last four
are numbers. The letters are not case-sensitive.}
}
\description{
This function helps you retrieve document metadata from the OIDA archives.
}
\details{
Documentation for the OIDA API can be found
\href{https://www.industrydocuments.ucsf.edu/wp-content/uploads/2022/02/IndustryDocumentsDataAPI_v5.pdf}{here} The function retrieves only the metadata for a document within the archives. Documents are stored as pdfs and must be downloaded individually within the main site.
}
\examples{
\dontrun{#Strategic Customer Group 2003 Business Plan
df <- get_oida_metadata(file_names = "frgg0230")
}
\dontrun{# get multiple files
file_names <- c("kylw0221", "frgg0230", "slwl0231")
df <- get_oida_metadata(file_names = file_names)
}

}
