% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oida_doc_index.R
\docType{data}
\name{oida_idx}
\alias{oida_idx}
\title{OIDA document metadata}
\format{
A tibble with 100 rows and 37 variables columns:
\describe{
\item{id}{id}
\item{tid}{tid}
\item{collection}{collection}
\item{collectioncode}{collectioncode}
\item{box}{box}
\item{availability}{availability}
\item{case}{case}
\item{area}{area}
\item{title}{title}
\item{author}{author}
\item{documentdate}{documentdate}
\item{type}{type}
\item{pages}{pages}
\item{brand}{brand}
\item{description}{description}
\item{file}{file}
\item{bates}{bates}
\item{batesmaster}{batesmaster}
\item{format}{format}
\item{dateproduced}{dateproduced}
\item{dateshipped}{dateshipped}
\item{dateaddeducsf}{dateaddeducsf}
\item{dateaddedindustry}{dateaddedindustry}
\item{datemodifiedindustry}{datemodifiedindustry}
\item{topic}{topic}
\item{mentioned}{mentioned}
\item{minnesotarequestnumber}{minnesotarequestnumber}
\item{requestnumber}{requestnumber}
\item{batesalternate}{batesalternate}
\item{othernumber}{othernumber}
\item{copied}{copied}
\item{cited}{cited}
\item{recipient}{recipient}
\item{privilegecode}{privilegecode}
\item{redacted}{redacted}
\item{access}{access}
\item{genre}{genre}
...
}
}
\source{
Additional information is available on the OIDA
\href{https://www.industrydocuments.ucsf.edu/opioids/research-tools/api/}{website}
}
\usage{
oida_idx
}
\description{
Documents housed digitally at OIDA contain metadata so that users can
find relevant documents. This dataset contains the first 100 results from
a search where the author was "glantz."
}
\details{
The OIDA search results are included here as a tibble. Investigators of the
archive can practice on a representative dataset within the R programming
environment.
}
\keyword{datasets}
